Fixed bug on add-courses with section-specific titles
Because sections with the same section-specific title as the course name are listed without a section-specific title header, if for some reason they weren't listed first in the sort, they could appear to be grouped in with sections with different titles. Now the header is only not shown if the section-specific titles match the course name AND that specific group is listed first.